Add files via upload
Results from using the model "all-MiniLM-L6-v2". These results contain json files  with the classification of each pair from texts in DATASET, the global results and a csv file with the comparison against the gold-standard. Due to the detailed information on similar sentences and the comparison, the json files of sentences and the json of the comparison are placed in the external repository...